PA-7: increment score in db when zombie killed, and decrease helth on attack

+++ core/src/com/mygdx/game/items/PlayerUpdateModel.java
@@ -3,13 +3,13 @@ package com.mygdx.game.items;
 public class PlayerUpdateModel {
     // Firebase needs properties to be public to work
     public String player;
-    public float health;
-    public Integer score;
+    public int health;
+    public int score;
 
     // Default constructor required for calls to DataSnapshot.getValue(User.class)
     public PlayerUpdateModel() {}
 
-    public PlayerUpdateModel(String player, float health, Integer score) {
+    public PlayerUpdateModel(String player, int health, int score) {
         this.player = player;
         this.health = health;
         this.score = score;
